Removed owner as a property of drafts; it is now passed when `paste()`ing the draft to Pastebin.
