+++ Pastes/AbstractPaste.php
@@ -3,7 +3,6 @@
 namespace Brush\Pastes {
 
 	use \Brush\Accounts\Developer;
-	use \Brush\Accounts\Account;
 	use \Brush\Pastes\Options\Format;
 	use \Brush\Pastes\Options\Visibility;
 	use \Brush\Exceptions\ValidationException;
@@ -41,13 +40,6 @@ namespace Brush\Pastes {
 		 * @var int
 		 */
 		private $visibility;
-
-		/**
-		 * The account that owns this paste.
-		 * This is required if the visibility of this paste is private.
-		 * @var \Brush\Accounts\Account
-		 */
-		private $owner;
 
 		/**
 		 * Retrieve the name of this paste.
@@ -114,31 +106,6 @@ namespace Brush\Pastes {
 		}
 
 		/**
-		 * Retrieve the account that owns this paste.
-		 * @return \Brush\Accounts\Account The account that owns this paste.
-		 */
-		public final function getOwner() {
-			return $this->owner;
-		}
-
-		/**
-		 * Find whether this paste has an owner set.
-		 * If it doesn't, it cannot have private visibility.
-		 * @return boolean True if this paste has an assigned owner; false otherwise.
-		 */
-		public final function hasOwner() {
-			return $this->getOwner() !== null;
-		}
-
-		/**
-		 * Set the account that owns this paste.
-		 * @param \Brush\Accounts\Account $owner The account that owns this paste.
-		 */
-		public final function setOwner(Account $owner) {
-			$this->owner = $owner;
-		}
-
-		/**
 		 * Initialise properties to defaults.
 		 */
 		protected function __construct() {
@@ -179,9 +146,6 @@ namespace Brush\Pastes {
 			$variables->set('api_paste_code', $this->getContent());
 			$variables->set('api_paste_private', $this->getVisibility());
 			$this->getFormat()->addTo($request);
-			if ($this->hasOwner()) {
-				$this->getOwner()->sign($request, $developer);
-			}
 		}
 	}
 }
